+++ src/main/kotlin/com/nylas/models/CreateWorkspaceRequest.kt
@@ -14,19 +14,19 @@ data class CreateWorkspaceRequest(
   val autoGroup: Boolean? = null,
   @Json(name = "policy_id")
   val policyId: String? = null,
-  @Json(name = "rules_ids")
-  val rulesIds: List<String>? = null,
+  @Json(name = "rule_ids")
+  val ruleIds: List<String>? = null,
 ) {
   data class Builder(private val name: String) {
     private var domain: String? = null
     private var autoGroup: Boolean? = null
     private var policyId: String? = null
-    private var rulesIds: List<String>? = null
+    private var ruleIds: List<String>? = null
 
     fun domain(domain: String?) = apply { this.domain = domain }
     fun autoGroup(autoGroup: Boolean?) = apply { this.autoGroup = autoGroup }
     fun policyId(policyId: String?) = apply { this.policyId = policyId }
-    fun rulesIds(rulesIds: List<String>?) = apply { this.rulesIds = rulesIds }
-    fun build() = CreateWorkspaceRequest(name, domain, autoGroup, policyId, rulesIds)
+    fun ruleIds(ruleIds: List<String>?) = apply { this.ruleIds = ruleIds }
+    fun build() = CreateWorkspaceRequest(name, domain, autoGroup, policyId, ruleIds)
   }
 }